+++ tagger/src/lib.rs
@@ -1,6 +1,8 @@
 mod components;
+mod divider;
 mod event;
 mod matrix;
+mod observer;
 mod ordpaths;
 mod picker;
 mod state;
@@ -8,5 +10,7 @@ mod tagger;
 mod terminal;
 mod utils;
 
+pub use divider::Divider;
+pub use observer::Observer;
 pub use picker::{Method, Picker};
 pub use tagger::Tagger;