+++ CHANGELOG.md
@@ -9,6 +9,12 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 -->
 
 ## [Unreleased]
+## [0.1.0-alpha26] - 2024-09-24
+
+- generate auto complete scripts
+- enhance tagger: add `tagger divide` subcommand
+- enhance tagger: add `tagger observe` subcommand
+
 ## [0.1.0-alpha25] - 2024-09-22
 
 - improve: show path and error if failed to load image